优化物料保存接口

+++ common/data/src/main/java/org/thingsboard/server/common/data/TSyncMaterialBom.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "t_sync_material_bom")
@@ -27,6 +28,6 @@ public class TSyncMaterialBom {
 
     @Column(name = "ratio")
     @ApiModelProperty("比例")
-    private Double ratio;
+    private BigDecimal ratio;
 
 }
